refactor(team): isolate mailbox delivery transport
