+++ packages/coding-agent/src/gjc-runtime/team-launch.ts
@@ -1,0 +1,7 @@
+import type { GjcTeamStartOptions } from "./team-runtime";
+export function withTeamLaunchTransport(
+	options: GjcTeamStartOptions,
+	mailboxDeliveryTransport: GjcTeamStartOptions["mailboxDeliveryTransport"],
+): GjcTeamStartOptions {
+	return { ...options, mailboxDeliveryTransport };
+}
