+++ packages/coding-agent/src/gjc-runtime/team-notify.ts
@@ -1,0 +1,17 @@
+import type {
+	GjcTeamMailboxDeliveryInput,
+	GjcTeamMailboxDeliveryResult,
+	GjcTeamMailboxDeliveryTransport,
+} from "./team-store";
+
+export async function deliverTeamMailboxMessage(
+	transport: GjcTeamMailboxDeliveryTransport | undefined,
+	input: GjcTeamMailboxDeliveryInput,
+): Promise<GjcTeamMailboxDeliveryResult | null> {
+	if (!transport) return null;
+	try {
+		return await transport.deliverMailboxMessage(input);
+	} catch {
+		return null;
+	}
+}
