+++ packages/coding-agent/src/gjc-runtime/team-workers.ts
@@ -1,0 +1,15 @@
+import type { GjcTeamConfig, GjcTeamWorker } from "./team-runtime";
+export interface GjcTeamWorkerLifecycleContext {
+	config: GjcTeamConfig;
+	worker: GjcTeamWorker;
+	cwd: string;
+	env: NodeJS.ProcessEnv;
+}
+export function workerLifecycleContext(
+	config: GjcTeamConfig,
+	worker: GjcTeamWorker,
+	cwd: string,
+	env: NodeJS.ProcessEnv,
+): GjcTeamWorkerLifecycleContext {
+	return { config, worker, cwd, env };
+}
